collect and store data to the database

+++ src/main/java/com/ppp/billing/repository/ContentTypeRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ppp.billing.model.ContentType;
 
 
-public interface ContentTypeRepository extends JpaRepository<ContentType, Long> {
+public interface ContentTypeRepository extends JpaRepository<ContentType, Integer> {
 
 	Optional<ContentType>  findByName(String name);
-	Optional<ContentType>  findById(long id );
+	Optional<ContentType>  findById(int id );
 
 }
